pyproject.toml
[project]
name = "brandmint"
version = "4.5.0"
description = "Unified brand creation orchestrator — 45 skills, visual assets, 23-artifact NotebookLM publishing with prose synthesis"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "requests>=2.28",
    "fal-client>=0.4.0",
]

[project.scripts]
brandmint = "brandmint.cli.app:main"
bm = "brandmint.cli.app:main"

[project.optional-dependencies]
dev = ["pytest", "ruff"]
publishing = ["notebooklm-py>=0.3.2"]
vision = [
    "Pillow>=10.0",
    "colorgram.py>=1.2",
    "imagehash>=4.3",
    "numpy>=1.24",
    "scikit-image>=0.21",
    "opencv-python-headless>=4.8",
    "scipy>=1.11",
]
embeddings = [
    "transformers>=4.35",
    "torch>=2.1",
    "sentence-transformers>=2.2",
    "faiss-cpu>=1.7",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["brandmint*"]
